Inc. commit.

+++ splitstab.sql
@@ -5,6 +5,6 @@ use mydb;
 create table if not exists splitstab (
     tdate date,
     symbol varchar(18),
-    post numeric(18,8),
-    pre numeric(18,8)
+    post integer,
+    pre integer
 );
